fmt: make fractional formatting use `u32`
I think when I originally wrote fractional formatting, I was thinking
about using precision values greater than `9`. But that never really
panned out. And I used signed values because I used signed values
everywhere.

I think this is the last change we need to be able to feasibly add
native `std::time::Duration` support.

+++ src/fmt/mod.rs
@@ -430,7 +430,7 @@ trait WriteExt: Write {
     fn write_fraction(
         &mut self,
         formatter: &FractionalFormatter,
-        n: impl Into<i64>,
+        n: impl Into<u32>,
     ) -> Result<(), Error> {
         self.write_fractional(&Fractional::new(formatter, n.into()))
     }
